requirements.txt
pytesseract~=0.3.10
pillow~=10.3.0
PyMuPDF~=1.24.2
spacy~=3.7.4 # https://spacy.io/usage/v3